last read metadata on data bus storage blobs

+++ Rebus.Tests/Contracts/DataBus/Factories/FileSystemDataBusStorageFactory.cs
@@ -12,10 +12,7 @@ namespace Rebus.Tests.Contracts.DataBus.Factories
 
         public FileSystemDataBusStorageFactory()
         {
-            if (Directory.Exists(DirectoryPath))
-            {
-                Directory.Delete(DirectoryPath, true);
-            }
+            CleanUpDirectory();
         }
 
         public IDataBusStorage Create()
@@ -27,6 +24,14 @@ namespace Rebus.Tests.Contracts.DataBus.Factories
 
         public void CleanUp()
         {
+            CleanUpDirectory();
+        }
+
+        static void CleanUpDirectory()
+        {
+            if (!Directory.Exists(DirectoryPath)) return;
+
+            Console.WriteLine($"Removing directory '{DirectoryPath}'");
             Directory.Delete(DirectoryPath, true);
         }
     }
